Remote View: DB functionality

+++ src/test/scala/test/package.scala
@@ -9,15 +9,20 @@ package object test {
   def cleanDirectory(dir: String): Boolean = cleanDirectory(new File(dir))
 
   def cleanDirectory(dir: File): Boolean = {
-    (for {
-      item <- dir.listFiles()
-    } yield {
-      if (item.isDirectory) {
-        cleanDirectory(item)
-      } else {
-        item.delete()
-      }
-    }).forall { x => x } && dir.delete()
+    if (dir == null) {
+      true
+    } else {
+      dir.mkdirs()
+      (for {
+        item <- dir.listFiles()
+      } yield {
+        if (item.isDirectory) {
+          cleanDirectory(item)
+        } else {
+          item.delete()
+        }
+      }).forall { x => x} && dir.delete()
+    }
   }
 
   def timed[T](x: => T): (Long, T) = {
